added snippets field to scrape config
will be used to define code snippets for libraries.
this is essential for a fast code generator, since a search over all possible programs satisfying given user requirements is not feasible.

for now the snippets will have to be provided by the user that configures the scraper. in future it would be interesting to consider parsing code snippets on documentation pages to automatically generate snippets.