implemented scrape caching
don't rescrape a lib if the scrape config did not change in a way that affects the scraping behavior. this type of caching is turned on by default but can be disabled to force a new scrape.

useful to quickly update handwritten code snippets for libraries without triggering a full scrape.