optimized receive-actions
receivers are now found using direct index lookup instead of datalog.
this required changing the model: a new abstract data-receiver concept was introduced to model the ability of a concrete concept to partake in datatype and value propagation in CFGs.

TODO: receives(-semantic) is no longer looked up at the result of a call-result => those edges must be added by call-(completion-)actions. NOT DONE YET!